pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dozor"
version = "1.0.0"
description = "Dozor - AI-first server monitoring agent with MCP integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Anatoly Koptev", email = "anatolykoptev@gmail.com"}
]
keywords = ["mcp", "monitoring", "server", "docker", "security", "ai"]

dependencies = [
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
sse = [
    "starlette>=0.35.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
]
all = [
    "dozor[sse,dev]",
]

[project.scripts]
dozor = "dozor.cli:main"
dozor-mcp = "dozor.mcp_server:main"

[project.urls]
Homepage = "https://github.com/anatolykoptev/dozor"
Repository = "https://github.com/anatolykoptev/dozor"
Issues = "https://github.com/anatolykoptev/dozor/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["dozor*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]
